Introduce cache.modify to clean the cache after cancelling a trip
After cancelling a trip the cached user object still retained a
reference to the trip that was cancelled, making it look like
the user was still booked on that trip. This didn't cause any
issues when using the application as a manual `refetchQueries`
was being triggered, which in turn pulled new user details in from
the network, and overwrote the old trip data in the cache. This
is a good oppoprtunity to demonstrate how the new `cache.modify()`
API works by removing the manual `refetchQueries` call and instead
just removing the cancelled trip directly from the user object
in the cache, after the trip was cancelled.

Note that to accomplish this we need to persist the user's `id`
(so it can be used when looking up the current user object in the
cache). This tutorial was setup to persist the user token in
`localStorage`, so we're now also storing the user `id` in
`localStorage`. A better option would be to adjust the schema so
we can get the current user `id` back in the trip cancellation
response (and thereby avoiding `localStorage`), but for now this
is the easiest option (and the user `id` is not really the focus
of this tutorial).